% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_printing.R
\name{num}
\alias{num}
\title{Display a number in Dollars}
\usage{
num(x, use_cents = FALSE, digits, decimals)
}
\arguments{
\item{x}{The numeric value to be conferted (must be length 1)}

\item{use_cents}{Should 2 digits to the right of the decimal be used (default = FALSE)}

\item{digits}{Specify the number of digits to use (includes left and right of decimal)}

\item{decimals}{Specifies the number of digits to the right of the decimal}
}
\description{
This function accepts a single number and converts it into dollars for display in
the monitoring reports.
}
