% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_printing.R
\name{round4display}
\alias{round4display}
\title{Round data.table Columns for Displaying}
\usage{
round4display(dt, var_list, decimals = 0)
}
\arguments{
\item{dt}{Data.table to be changed}

\item{var_list}{A vector of variable names (as characters)}

\item{decimals}{The number of decimals to display}
}
\description{
This function replaces the numbers in a data.table with rounded versions that
are limited to a specified number of decimal places.  The changes are made
by reference.
}
